Prelim styling of charts
